backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
joblib==1.4.2
mpmath==1.3.0
nltk==3.9.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pgvector
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pandas
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.17.0
sqlparse==0.5.2
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
fastapi
uvicorn
prisma
python-dotenv==1.0.1
langchain
langchain-community
sentence-transformers
faiss-cpu
smolagents
tiktoken
openai
weaviate-client
pydantic